# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/form/admin": {
      "post": {
        "description": "지원 폼을 생성합니다. (응답 생성 X)",
        "operationId": "FormPrivateController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFormDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FormDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "지원 폼 생성",
        "tags": [
          "FormPrivate"
        ]
      },
      "get": {
        "operationId": "FormPrivateController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FormDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "지원 폼 모두 가져오기",
        "tags": [
          "FormPrivate"
        ]
      }
    },
    "/form/admin/{id}": {
      "patch": {
        "operationId": "FormPrivateController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFormDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "$ref": "#/components/schemas/FormDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "특정 지원 폼 수정",
        "tags": [
          "FormPrivate"
        ]
      },
      "delete": {
        "operationId": "FormPrivateController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "type": "string",
                          "example": "ok"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "지원 폼 삭제",
        "tags": [
          "FormPrivate"
        ]
      },
      "get": {
        "operationId": "FormPrivateController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "$ref": "#/components/schemas/FormDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "지원 폼 데이터 가져오기",
        "tags": [
          "FormPrivate"
        ]
      }
    },
    "/form/admin/{id}/responses": {
      "get": {
        "operationId": "FormPrivateController_findAllResponses",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FormResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "특정 지원 폼의 모든 응답 가져오기",
        "tags": [
          "FormPrivate"
        ]
      }
    },
    "/form/admin/responses/{responseId}": {
      "get": {
        "operationId": "FormPrivateController_findOneResponse",
        "parameters": [
          {
            "name": "responseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "$ref": "#/components/schemas/FormResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "특정 지원 폼의 특정 응답 가져오기",
        "tags": [
          "FormPrivate"
        ]
      }
    },
    "/form/admin/{id}/activate": {
      "post": {
        "operationId": "FormPrivateController_activateForm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "$ref": "#/components/schemas/FormDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "지원 폼 활성화",
        "tags": [
          "FormPrivate"
        ]
      }
    },
    "/form/admin/{id}/deactivate": {
      "post": {
        "operationId": "FormPrivateController_deactivateForm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponseDto"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "example": 200
                        },
                        "message": {
                          "type": "string"
                        },
                        "responseAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "data": {
                          "$ref": "#/components/schemas/FormDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "지원 폼 비활성화",
        "tags": [
          "FormPrivate"
        ]
      }
    },
...
  },
  "info": {
    "title": "TAPIE API",
    "description": "TAPIE System API",
    "version": "0.0.8",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8877",
      "description": "Local Development"
    },
    {
      "url": "https://tapie-api-development.vport.dev",
      "description": "Remote Development"
    },
    {
      "url": "https://api.tapie.kr",
      "description": "Remote Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FormDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "폼 ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "폼 이름",
            "example": "Sample Form"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string",
            "description": "시작 날짜"
          },
          "endsAt": {
            "format": "date-time",
            "type": "string",
            "description": "종료 날짜"
          },
          "active": {
            "type": "boolean",
            "description": "활성화 여부"
          }
        },
        "required": [
          "id",
          "name",
          "startsAt",
          "endsAt",
          "active"
        ]
      },
      "FormResponseDto": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "응답 UUID"
          },
          "formId": {
            "type": "number",
            "description": "폼 ID"
          },
          "memberUUID": {
            "type": "string",
            "description": "멤버 UUID"
          },
          "portfolioAssetUUID": {
            "type": "string",
            "description": "포트폴리오 UUID"
          },
          "name": {
            "type": "string",
            "description": "이름"
          },
          "studentId": {
            "type": "string",
            "description": "학번"
          },
          "googleEmail": {
            "type": "string",
            "description": "구글 이메일"
          },
          "unit": {
            "type": "string",
            "description": "유닛",
            "enum": [
              "DEVELOPER",
              "DESIGNER",
              "EXTERNAL"
            ]
          },
          "phoneNumber": {
            "type": "string",
            "description": "전화번호"
          },
          "introduction": {
            "type": "string",
            "description": "소개"
          },
          "motivation": {
            "type": "string",
            "description": "동기"
          },
          "expectedActivities": {
            "type": "string",
            "description": "기대 활동"
          },
          "reasonToChoose": {
            "type": "string",
            "description": "선택 이유"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "생성일"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "수정일"
          },
          "submitted": {
            "type": "boolean",
            "description": "제출 여부"
          }
        },
        "required": [
          "uuid",
          "formId",
          "memberUUID",
          "portfolioAssetUUID",
          "name",
          "studentId",
          "googleEmail",
          "unit",
          "phoneNumber",
          "introduction",
          "motivation",
          "expectedActivities",
          "reasonToChoose",
          "createdAt",
          "updatedAt",
          "submitted"
        ]
      },
      "CreateFormDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "폼 이름",
            "example": "Sample Form"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string",
            "description": "시작 날짜"
          },
          "endsAt": {
            "format": "date-time",
            "type": "string",
            "description": "종료 날짜"
          },
          "active": {
            "type": "boolean",
            "description": "활성화 여부"
          }
        },
        "required": [
          "name",
          "startsAt",
          "endsAt",
          "active"
        ]
      },
      "UpdateFormDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "폼 이름",
            "example": "Sample Form"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string",
            "description": "시작 날짜"
          },
          "endsAt": {
            "format": "date-time",
            "type": "string",
            "description": "종료 날짜"
          },
          "active": {
            "type": "boolean",
            "description": "활성화 여부"
          }
        }
      },
...
    }
  }
}